+++ dbt/adapters/glue/credentials.py
@@ -1,8 +1,7 @@
 from dataclasses import dataclass
 from typing import Optional
-from dbt.adapters.base import Credentials
-import dbt.exceptions
-
+from dbt.adapters.contracts.connection import Credentials
+from dbt_common.exceptions import DbtRuntimeError
 
 @dataclass
 class GlueCredentials(Credentials):
@@ -56,7 +55,7 @@ class GlueCredentials(Credentials):
     def __post_init__(self):
         # spark classifies database and schema as the same thing
         if self.database is not None and self.database != self.schema:
-            raise dbt.exceptions.DbtRuntimeError(
+            raise DbtRuntimeError(
                 f"    schema: {self.schema} \n"
                 f"    database: {self.database} \n"
                 f"On Spark, database must be omitted or have the same value as"