+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-## next version
+## New versio
+- Fix session provisioning timeout and delay handling
+
+## v1.8.1
 - Fix typo in README.md
 - Fix unit test failure caused by moto 5 upgrade
 - Fix pagination bug when listing glue databases and tables